% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{cleantibble}
\alias{cleantibble}
\title{Tibble cleaning}
\usage{
cleantibble(tibble_list)
}
\arguments{
\item{tibble_list}{a list generated from "tibbleconvert" function.}
}
\value{
returns a list of tibbles.
}
\description{
Tibble cleaning
}
