% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{compact}
\alias{compact}
\title{Selection of variables}
\usage{
compact(clean_tibble)
}
\arguments{
\item{clean_tibble}{takes a list of clean tibbles.}
}
\value{
returns a a list of tibbles.
}
\description{
Selection of variables
}
