% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{folder_to_df}
\alias{folder_to_df}
\title{Entire transform, folder edition}
\usage{
folder_to_df(folder_path)
}
\arguments{
\item{folder_path}{takes a folder file path. Transforms all of the XML files.}
}
\value{
returns a single tibble of one row per XML input file in folder.
}
\description{
Entire transform, folder edition
}
