% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{import_split}
\alias{import_split}
\title{Importing and splitting}
\usage{
import_split(xml_file_path)
}
\arguments{
\item{xml_file_path}{file path to a single xml file generated på PimSOFT. Mandatory recinfo, calc, meanperf and roi info.}
}
\value{
returns a list of data frames. One long column of raw data.
}
\description{
Importing and splitting
}
