% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_xml_folder.R
\name{process_xml_folder}
\alias{process_xml_folder}
\title{process xml folder to one tibble data frame}
\usage{
process_xml_folder(folder_path)
}
\arguments{
\item{folder_path}{takes path to folder containing xml files}
}
\value{
returns a tibble of all data in folder
}
\description{
process xml folder to one tibble data frame
}
