% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_xml_folder.R
\name{process_xml_single}
\alias{process_xml_single}
\title{process xml file to one tibble data frame}
\usage{
process_xml_single(xml_path)
}
\arguments{
\item{xml_path}{takes path to folder containing xml files}
}
\value{
returns a tibble of onw row
}
\description{
process xml file to one tibble data frame
}
