% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{singlerow}
\alias{singlerow}
\title{Transforms to singlerow}
\usage{
singlerow(compact_tibble)
}
\arguments{
\item{compact_tibble}{takes a list of compact tibbles from "compact" function.}
}
\value{
returns a single tibble of one row per XML input file.
}
\description{
Transforms to singlerow
}
