% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlerow_df_compile.R
\name{singlerow_df_compile}
\alias{singlerow_df_compile}
\title{singlerow_df_compile}
\usage{
singlerow_df_compile(xml_folder_file_path)
}
\arguments{
\item{xml_folder_file_path}{}
}
\value{
A dataframe with each XML file as an observation(row)
}
\description{
An addon to function "xml_to_singlerow" which compiles each observation(xml file) in to an easy to read data frame format.
}
