% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{tibbleconvert}
\alias{tibbleconvert}
\title{Tibble converting}
\usage{
tibbleconvert(list)
}
\arguments{
\item{list}{a list generated from "import_split" function}
}
\value{
returns a list of tibbles and necessary variable data.
}
\description{
Tibble converting
}
