% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles_to_singlerow.R
\name{tibbles_to_singlerow}
\alias{tibbles_to_singlerow}
\title{turn tibbles in to single row data}
\usage{
tibbles_to_singlerow(tibbles_list)
}
\arguments{
\item{tibbles_list}{a list of 4 tibbles generated by xml_to_tibbles function.}
}
\value{
returns 1 tibble with one singlerow.
}
\description{
*Applicable to microvasc research group, Danderyds Sjukhus workflow*. Takes tibbles from "xml_to_tibbles" and generates 4 tibbles of single row data.
}
