% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1.tsv.R
\name{tsv_convert}
\alias{tsv_convert}
\title{Convert TSV File}
\usage{
tsv_convert(tsv_file)
}
\arguments{
\item{tsv_file}{A data frame representing the TSV file.}
}
\value{
A list with data frame with cleaned column names and converted units.
}
\description{
This function cleans column names and converts units in a TSV file.
}
\examples{
tsv_convert(my_tsv_file)

}
