% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2.R
\name{xml_to_df}
\alias{xml_to_df}
\title{Entire transform}
\usage{
xml_to_df(xml_file_path)
}
\arguments{
\item{xml_file_path}{takes a XML file path}
}
\value{
returns a single tibble of one row per XML input file.
}
\description{
Entire transform
}
