% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfunctions.R
\name{xml_to_singlerow}
\alias{xml_to_singlerow}
\title{xml_to_singlerow}
\usage{
xml_to_singlerow(
  xml_file_path,
  recordinginfo = FALSE,
  calculations = FALSE,
  meanperfusion = FALSE,
  changeperROI = FALSE
)
}
\arguments{
\item{xml_file_path}{file path to xml file in working directory}

\item{recordinginfo}{is not used at the moment}

\item{calculations}{is not used at the moment}

\item{meanperfusion}{is not used at the moment}

\item{changeperROI}{is not used at the moment}
}
\value{
Returns a single row tibble dataframe.
}
\description{
*Applicable to microvasc research group, Danderyds Sjukhus workflow*. Takes a PimSOFT report in XML format and generates a single row of variables suited for implementation into a data frame. It is suited for reports generated with "Recording info" and "Mean perfusion" boxes ticked.
}
