% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_split.R
\name{xml_to_split}
\alias{xml_to_split}
\title{xml_to_split}
\usage{
xml_to_split(
  xml_file_path,
  recordinginfo = FALSE,
  calculations = FALSE,
  meanperfusion = FALSE,
  changeperROI = FALSE
)
}
\arguments{
\item{changeperROI}{}
}
\description{
xml_to_split
}
