% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_tibbles.R
\name{xml_to_tibbles}
\alias{xml_to_tibbles}
\title{turning XML to seperate tibbles}
\usage{
xml_to_tibbles(xml_file_path)
}
\arguments{
\item{xml_file_path}{file path to xml file in working directory}
}
\value{
returns 4 tibbles in a list with clean data.
}
\description{
*Applicable to microvasc research group, Danderyds Sjukhus workflow*. Takes a PimSOFT report in XML format and generates clean data frames in tibble format. One data frame per specified report category. Recinfo, calc, meanperf, roi mandatory.
}
