% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importfunctions.R
\name{singlerow_df_compile}
\alias{singlerow_df_compile}
\title{singlerow_df_compile}
\usage{
singlerow_df_compile(
  xml_folder_file_path,
  recordinginfo = FALSE,
  calculations = FALSE,
  meanperfusion = FALSE,
  changeperROI = FALSE
)
}
\arguments{
\item{xml_folder_file_path}{file path to folder that needs to be compiled.}

\item{recordinginfo}{is not used at the moment}

\item{calculations}{is not used at the moment}

\item{meanperfusion}{is not used at the moment}

\item{changeperROI}{is not used at the moment}
}
\value{
Returns a tibble dataframe.
}
\description{
*Applicable to microvasc research group, Danderyds Sjukhus workflow*. Takes a *FOLDER* of PimSOFT report in XML format and generates a single row of variables suited for implementation into a data frame. It is suited for reports generated with "Recording info" and "Mean perfusion" boxes ticked.
}
