% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_singlerow.R
\name{xml_to_singlerow}
\alias{xml_to_singlerow}
\title{xml_to_singlerow}
\usage{
xml_to_singlerow(xml_file_path)
}
\arguments{
\item{xml_file_path}{file path to xml file in working directory}
}
\value{
returns a single row of variables of xml file
}
\description{
*Applicable to microvasc research group, Danderyds Sjukhus workflow*. Takes a PimSOFT report in XML format and generates a single row of variables suited for implementation into a data frame. It is suited for reports generated with "Recording info" and "Mean perfusion" boxes ticked.
}
